# openapi.yaml
openapi: 3.0.0
info:
  title: Food Recipes API
  version: 1.0.0
  description: API for managing food recipes
servers:
  - url: http://localhost:3000
paths:
  /recipes:
    get:
      summary: List all recipes
      responses:
        '200':
          description: A list of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
              example:
                - id: "tx001"
                  name: "Texas-Style Beef Brisket"
                  ingredients:
                    - "1 whole beef brisket (10-12 pounds)"
                    - "1/4 cup kosher salt"
                    - "1/4 cup black pepper"
                    - "2 tablespoons garlic powder"
                    - "2 tablespoons onion powder"
                    - "1 tablespoon cayenne pepper"
                  instructions: "1. Mix all spices together to create a rub. 2. Trim excess fat from brisket, leaving about 1/4 inch fat cap. 3. Apply rub generously to all sides of brisket. 4. Let sit for 1 hour at room temperature. 5. Prepare smoker to 225°F using oak or hickory wood. 6. Smoke brisket fat side up for 6 hours. 7. Wrap in butcher paper and continue cooking until internal temperature reaches 203°F (about 6 more hours). 8. Rest for 1-2 hours before slicing against the grain."
                  cuisine: "Texas BBQ"
                  prepTime: 60
                  cookTime: 720
    post:
      summary: Add a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
            example:
              id: "tx001"
              name: "Texas-Style Beef Brisket"
              ingredients:
                - "1 whole beef brisket (10-12 pounds)"
                - "1/4 cup kosher salt"
                - "1/4 cup black pepper"
                - "2 tablespoons garlic powder"
                - "2 tablespoons onion powder"
                - "1 tablespoon cayenne pepper"
              instructions: "1. Mix all spices together to create a rub. 2. Trim excess fat from brisket, leaving about 1/4 inch fat cap. 3. Apply rub generously to all sides of brisket. 4. Let sit for 1 hour at room temperature. 5. Prepare smoker to 225°F using oak or hickory wood. 6. Smoke brisket fat side up for 6 hours. 7. Wrap in butcher paper and continue cooking until internal temperature reaches 203°F (about 6 more hours). 8. Rest for 1-2 hours before slicing against the grain."
              cuisine: "Texas BBQ"
              prepTime: 60
              cookTime: 720
  /recipes/{id}:
    get:
      summary: Get a recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
              example:
                id: "tx001"
                name: "Texas-Style Beef Brisket"
                ingredients:
                  - "1 whole beef brisket (10-12 pounds)"
                  - "1/4 cup kosher salt"
                  - "1/4 cup black pepper"
                  - "2 tablespoons garlic powder"
                  - "2 tablespoons onion powder"
                  - "1 tablespoon cayenne pepper"
                instructions: "1. Mix all spices together to create a rub. 2. Trim excess fat from brisket, leaving about 1/4 inch fat cap. 3. Apply rub generously to all sides of brisket. 4. Let sit for 1 hour at room temperature. 5. Prepare smoker to 225°F using oak or hickory wood. 6. Smoke brisket fat side up for 6 hours. 7. Wrap in butcher paper and continue cooking until internal temperature reaches 203°F (about 6 more hours). 8. Rest for 1-2 hours before slicing against the grain."
                cuisine: "Texas BBQ"
                prepTime: 60
                cookTime: 720
        '404':
          description: Recipe not found
    put:
      summary: Update a recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
      responses:
        '200':
          description: Recipe updated
        '404':
          description: Recipe not found
    delete:
      summary: Delete a recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Recipe deleted
        '404':
          description: Recipe not found
components:
  schemas:
    Recipe:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the recipe
        name:
          type: string
        ingredients:
          type: array
          items:
            type: string
        instructions:
          type: string
        cuisine:
          type: string
          enum: ["Tex-Mex", "Texas BBQ", "Texas Chili", "Southern Texas", "Cowboy/Ranch-style", "Gulf Coast"]
          description: Type of Texas cuisine
        prepTime:
          type: integer
          description: Preparation time in minutes
        cookTime:
          type: integer
          description: Cooking time in minutes
      required:
        - id
        - name
        - ingredients
        - instructions